+++ database/migrations/2017_03_12_091611_create_product_images_table.php
@@ -14,8 +14,9 @@ class CreateProductImagesTable extends Migration
 	public function up()
 	{
 		Schema::create('product_images', function (Blueprint $table) {
-			$table->increments('id');
-			$table->integer('id_product');
+			$table->bigIncrements('id');
+			// $table->integer('id_product');
+			$table->unsignedBigInteger('id_product');
 			$table->foreign('id_product')->references('id')->on('products');
 			$table->text('mime');
 			// $table->longText('image')->unique(); // --> for sqlite
