+++ database/migrations/2017_03_14_001723_create_products_table.php
@@ -14,11 +14,13 @@ class CreateProductsTable extends Migration
     public function up()
     {
         Schema::create('products', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('id_user');
+            $table->bigIncrements('id');
+            // $table->integer('id_user');
+            $table->unsignedBigInteger('id_user');
             $table->foreign('id_user')->references('id')->on('users');   // table name itself
             $table->string('product')->unique();
-            $table->integer('id_category');
+            // $table->integer('id_category');
+            $table->unsignedBigInteger('id_category');
             $table->foreign('id_category')->references('id')->on('categories');
             $table->float('retail', 8, 2);
             $table->float('commission', 8, 2);
