+++ database/migrations/2017_04_09_080258_create_product_categories_table.php
@@ -14,8 +14,9 @@ class CreateProductCategoriesTable extends Migration
 	public function up()
 	{
 		Schema::create('product_categories', function (Blueprint $table) {
-			$table->increments('id');
-            $table->integer('id_user');
+			$table->bigIncrements('id');
+            // $table->integer('id_user');
+            $table->unsignedBigInteger('id_user');
             $table->foreign('id_user')->references('id')->on('users')->onUpdate('cascade');   // table name itself
             $table->string('product_category')->unique();
             $table->boolean('active')->nullable();
