+++ database/migrations/2017_04_10_022558_create_sales_table.php
@@ -9,8 +9,9 @@ class CreateSalesTable extends Migration
 	public function up()
 	{
 		Schema::create('sales', function (Blueprint $table) {
-			$table->increments('id');
-			$table->integer('id_user');
+			$table->bigIncrements('id');
+			// $table->integer('id_user');
+			$table->unsignedBigInteger('id_user');
 			$table->foreign('id_user')->references('id')->on('users');
 			$table->date('date_sale');
 			$table->softDeletes();
