+++ database/migrations/2017_04_10_022855_create_sales_items_table.php
@@ -14,10 +14,12 @@ class CreateSalesItemsTable extends Migration
     public function up()
     {
         Schema::create('sales_items', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('id_sales');
+            $table->bigIncrements('id');
+            // $table->integer('id_sales');
+            $table->unsignedBigInteger('id_sales');
             $table->foreign('id_sales')->references('id')->on('sales');
-            $table->integer('id_product');
+            // $table->integer('id_product');
+            $table->unsignedBigInteger('id_product');
             $table->foreign('id_product')->references('id')->on('products');
             $table->float('commission', 8, 2);
             $table->float('retail', 8, 2);
