+++ database/migrations/2017_04_29_145506_create_payments_table.php
@@ -14,10 +14,12 @@ class CreatePaymentsTable extends Migration
     public function up()
     {
         Schema::create('payments', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('id_sales');
+            $table->bigIncrements('id');
+            // $table->integer('id_sales');
+            $table->unsignedBigInteger('id_sales');
             $table->foreign('id_sales')->references('id')->on('sales');
-            $table->integer('id_bank');
+            // $table->integer('id_bank');
+            $table->unsignedBigInteger('id_bank');
             $table->foreign('id_bank')->references('id')->on('banks');
             $table->date('date_payment');
             $table->float('amount', 8, 2);
