+++ database/migrations/2017_05_03_012957_create_sales_taxes_table.php
@@ -14,10 +14,12 @@ class CreateSalesTaxesTable extends Migration
     public function up()
     {
         Schema::create('sales_taxes', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('id_sales');
+            $table->bigIncrements('id');
+            // $table->integer('id_sales');
+            $table->unsignedBigInteger('id_sales');
             $table->foreign('id_sales')->references('id')->on('sales');
-            $table->integer('id_tax');
+            // $table->integer('id_tax');
+            $table->unsignedBigInteger('id_tax');
             $table->foreign('id_tax')->references('id')->on('taxes');
             $table->softDeletes();
             $table->timestamps();
